[FEAT] (AreaEditor.tsx, AreaEditorBoard.tsx, AreaEditorToolbar.tsx, ServiceCardItem.tsx, useAreaEditor.ts, AreaEditor.module.css): implement area editor with drag-and-drop functionality, service management, and UI enhancements
